refactor(main): extract npm toolchain into npm_toolchain (main.py god-file slice R3)
